refactor(net): split authoritative route responsibilities
Replace the monolithic route module with cohesive FIB, lifecycle, transaction, validation, and typed-model modules behind a narrow facade. Route mutations now pass through FibEditor and projection plans so affected interfaces, rollback state, and Linux-facing notifications cannot drift across call sites.

Model transient IPv4 output decisions and source-selection policy explicitly. This aligns explicit-OIF fallback, ingress broadcast, multicast gateway handling, lenient route dumps, link lifecycle notifications, and IPv6 no-source errors with Linux 6.6 semantics.

Add rtnetlink regression coverage for non-persistent limited broadcast results and gateway-preserving multicast lookups.
